feat(logical): allow logical expressions to coerce values into required types
In previous versions, logical expressions attempted to evaluate the parameters to the required types
and in case any parameter evaluated to an invalid type, an TypeError would be thrown. For example,
an $if expression would require its conditionExpression to be evaluated to a Boolean. This
modification allows logical expressions to coerce received parameters into their required types
instead of throwing TypeError

BREAKING CHANGE: Logical expressions no longer throw TypeError for parameters that do not satisfy
type requirements

re #2